requirements.txt
binance-connector~=3.3.1
pandas~=2.1.0
seaborn~=0.12.2
matplotlib~=3.7.2
python-decouple~=3.8
requests~=2.31.0